Add util.DeferredSignal and a couple other performance tweaks.
Removing ~7000 songs from the library in the search browser took 500 seconds
on my system. A little poking around showed that the PlaylistModel was firing
a signal after every individual song was removed. DeferredSignal allows calls
like this to be pushed into the idle loop and handled after the storm of
signals has passed, which is fine for most UI events. This, in addition to
a small change in the album browser, brought the total time to 4.1 seconds.

DeferredSignal probably needs unit tests, although I'm not sure how to test it
yet. It should also be used in more places.