Updated brainz.py. Closes issue 136. Closes issue 319.
This is close to a total overhaul. It requires musicbrainz2, the new
pure-Python version, but that library is in much better shape than the RDF one
so I feel it's OK to depend on it. This version writes 'labelid' tags, but
those will stop being special in a later version of QL, after MusicBrainz NGS
becomes popular.