gstbe: fix a deadlock when seeking during a gapless transition.
The about-to-finish handler is called with the stream lock
held. If we schedule a song change in the mainloop and wait
for the result it could happen that a seek event got scheduled
right inbetween from a previous action, which results in a deadlock.
(seek waits for the stream lock, the signal handler which owns it
waits for the main loop to get free)
Fix by adding a timeout to MainRunner.call which raises if after
the timout has passed the function hasn't been scheduled yet and
use it to abort the about-to-finish handling.
This results in the gapless transition being cancelled and
we'll get a EOS later on instead.